Better location for new function.
